Introduce AppBarHeader for the AppBar & Toolbar. Introduce snapshot testing - but we bypass the MaterialUI HoC for the shallow tests, as otherwise we don't see the component rendering.
