+++ src/modules/signup/pages/CreateOrganizationPage.jsx
@@ -1,48 +1,15 @@
 import React from "react";
-import { withRouter, Link } from "react-router-dom";
-import { withStyles } from "@material-ui/core/styles";
-import AppBar from "@material-ui/core/AppBar";
-import Toolbar from "@material-ui/core/Toolbar";
-import IconButton from "@material-ui/core/IconButton";
-import ArrowBack from "@material-ui/icons/ArrowBack";
-import Typography from "@material-ui/core/Typography";
+import { withRouter } from "react-router-dom";
 import Grid from "@material-ui/core/Grid";
 
 import { Page } from "modules/layout/components";
+import AppBarHeader from "components/AppBarHeader";
 
 import CreateOrganizationFlow from "../components/CreateOrganizationFlow";
 
-const styles = theme => ({
-  titleText: {
-    [theme.breakpoints.up("sm")]: {
-      margin: "0 auto"
-    }
-  }
-});
-
-export const CreateOrganizationPage = ({ classes, history }) => (
-  <div>
-    <AppBar position="fixed">
-      <Toolbar>
-        <IconButton
-          color="inherit"
-          aria-label="go back"
-          component={Link}
-          to="/signin"
-        >
-          <ArrowBack />
-        </IconButton>
-        <Typography
-          className={classes.titleText}
-          variant="h6"
-          color="inherit"
-          noWrap
-        >
-          {"Set up Boxwise"}
-        </Typography>
-      </Toolbar>
-    </AppBar>
-    <Toolbar />
+export const CreateOrganizationPage = ({ history }) => (
+  <React.Fragment>
+    <AppBarHeader backUrl="/signin" title="Set up Boxwise" />
     <Page>
       <Grid container spacing={24} justify="center" alignItems="center">
         <Grid item xs={12} md={4}>
@@ -56,7 +23,7 @@ export const CreateOrganizationPage = ({ classes, history }) => (
         </Grid>
       </Grid>
     </Page>
-  </div>
+  </React.Fragment>
 );
 
-export default withStyles(styles)(withRouter(CreateOrganizationPage));
+export default withRouter(CreateOrganizationPage);
